feat: make user confirm delete

+++ TravelPlanner/src/Components/Mypage/MyReviews/MyReviews.jsx
@@ -70,6 +70,10 @@ function MyReviews(props) {
   };
 
   const handleDelete = (reviewID) => {
+    const confimdDelete = window.confirm("Er du sikker på at du vil slette?");
+    if (!confimdDelete) {
+      return;
+    }
     try {
       API.delete(`/review/delete/${reviewID}`);
       console.log("Review has been deleted", reviewID);
